rename s3 routes to r2
