+++ worker/api/index.ts
@@ -2,7 +2,7 @@ import { Hono } from 'hono';
 import { zValidator } from '@hono/zod-validator';
 import { z } from 'zod';
 import authRoutes from './auth';
-import s3Routes from './s3';
+import r2Routes from './r2';
 import fileRoutes from './file';
 import { isBupt } from '../utils';
 import { byrdocs_login } from '@byrdocs/bupt-auth';
@@ -13,7 +13,7 @@ export default new Hono<{
 }>()
     .get('/ping', (c) => c.text('pong'))
     .route('/auth', authRoutes)
-    .route('/s3', s3Routes)
+    .route('/r2', r2Routes)
     .route('/file', fileRoutes)
     .get('/rank', async (c) => {
         const token = c.req.query('token');
